ml/requirements.txt
# PRISM ML Module - Production Dependencies
# Phase 4: Machine Learning Foundation
# Python 3.12+ Compatible Versions
# Last Updated: November 2025

# Core ML Libraries (Python 3.12 compatible)
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0

# Model Persistence & Serialization
joblib>=1.3.2,<2.0.0

# TensorFlow.js Export (for browser deployment - Phase 5)
# Note: Disabled for Python 3.12 Windows compatibility
# Will use ONNX or alternative export in Phase 5
# tensorflowjs>=4.17.0,<5.0.0
# tensorflow>=2.15.0,<3.0.0

# URL Analysis & Feature Extraction
urllib3>=2.1.0,<3.0.0
tldextract>=5.0.0,<6.0.0
validators>=0.22.0,<1.0.0

# Data Processing
scipy>=1.11.0,<2.0.0
imbalanced-learn>=0.12.0,<1.0.0

# Visualization & Reporting
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0

# Configuration & Environment
python-dotenv>=1.0.0,<2.0.0

# Logging & Monitoring
colorlog>=6.8.0,<7.0.0

# Development & Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
black>=24.0.0,<25.0.0
flake8>=7.0.0,<8.0.0
mypy>=1.8.0,<2.0.0

# Performance & Optimization
numba>=0.59.0,<1.0.0

# Additional Utilities
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
